+++ PowerShellForFOD/PowerShellForFOD.psm1
@@ -2,12 +2,12 @@
 # Above needs to remain the first line to import Classes
 # remove the comment when using classes
 
-#requires -Version 2
-#Get public and private function definition files.
+# requires -Version 2
+# Get public and private function definition files.
 $Public = @( Get-ChildItem -Path $PSScriptRoot\Public\*.ps1 -Recurse -ErrorAction SilentlyContinue )
 $Private = @( Get-ChildItem -Path $PSScriptRoot\Private\*.ps1 -Recurse -ErrorAction SilentlyContinue )
 
-#Dot source the files
+# Dot source the files
 Foreach ($import in @($Public + $Private)) {
     Try {
         . $import.fullname
@@ -17,17 +17,21 @@ Foreach ($import in @($Public + $Private)) {
     }
 }
 
-#Create / Read config
+# Create / Read config
 $script:_PS4FODXmlpath = Get-FODConfigPath
 if(-not (Test-Path -Path $script:_PS4FODXmlpath -ErrorAction SilentlyContinue))
 {
     Try
     {
         Write-Warning "Did not find config file $($script:_PS4FODXmlpath), attempting to create"
-        [pscustomobject]@{
+        [PSCustomObject]@{
             ApiUri = $null
+            GrantType = $null
+            Scope = $null
+            Credential = $null
             Token = $null
             Proxy = $null
+            ForceToken = $False
             ForceVerbose = $False
         } | Export-Clixml -Path $($script:_PS4FODXmlpath) -Force -ErrorAction Stop
     }
@@ -37,10 +41,10 @@ if(-not (Test-Path -Path $script:_PS4FODXmlpath -ErrorAction SilentlyContinue))
     }
 }
 
-#Initialize the config variable.
+# Initialize the config variable.
 Try
 {
-    #Import the config
+    # Import the config
     $PS4FOD = $null
     $PS4FOD = Get-FODConfig -Source PS4FOD.xml -ErrorAction Stop
 }